Update project structure.
