+++ src/components/ClientDetailPage/GeneralTable.tsx
