+++ src/components/ClientDetailPage/assets/ContributionWithdrawalsTable.tsx
@@ -1,8 +1,9 @@
 import React, { PureComponent } from 'react';
-import { Icon, Popconfirm, Table } from 'antd';
-import { InnerTableContainer, DivideLine, HeaderTitleTable, TextTitle } from '../styled';
+import {Icon, Popconfirm, Table} from 'antd';
+import { InnerTableContainer, HeaderTitleTable, TextTitle, DivideLine } from '../../../pages/client/styled';
+import { TweenOneGroup } from 'rc-tween-one';
 
-class CoverDetailsTable extends PureComponent {
+class ContributionWithdrawalsTable extends PureComponent {
   public state = {
     dataSource: [
       {
@@ -26,23 +27,16 @@ class CoverDetailsTable extends PureComponent {
         from: 'start',
         to: 'End',
       },
-      {
-        key: '3',
-        type: 'Custom',
-        value: 15000,
-        from: 'start',
-        to: 'End',
-      },
     ],
-    count: 4,
+    count: 3,
   };
 
   public columns = [
     {
       title: '',
       key: 'operation',
-      className: 'operation',
       width: 18,
+      className: 'operation',
       render: (text: any, record: any) =>
         this.state.dataSource.length >= 1 ? (
           <Popconfirm title="Sure to delete?" onConfirm={() => this.handleDelete(record.key)}>
@@ -74,6 +68,42 @@ class CoverDetailsTable extends PureComponent {
       width: 120,
     },
   ];
+
+  public enterAnim = [
+    {
+      opacity: 0,
+      x: 30,
+      duration: 0,
+      backgroundColor: '#fffeee',
+    },
+    {
+      duration: 200,
+      type: 'from',
+      delay: 250,
+      ease: 'easeOutQuad',
+    },
+    {
+      opacity: 1,
+      x: 0,
+      duration: 250,
+      ease: 'easeOutQuad',
+    },
+    { delay: 1000, backgroundColor: 'initial' },
+  ];
+  public leaveAnim = [{ duration: 250, opacity: 0 }, { duration: 200, ease: 'easeOutQuad' }];
+
+  public animTag = ($props: any) => {
+    return (
+      <TweenOneGroup
+        component="tbody"
+        enter={this.enterAnim}
+        leave={this.leaveAnim}
+        appear={false}
+        exclusive
+        {...$props}
+      />
+    );
+  }
 
   public handleDelete = (key: string) => {
     const dataSource = [...this.state.dataSource];
@@ -108,18 +138,18 @@ class CoverDetailsTable extends PureComponent {
         }),
       };
     });
-
+    const components = { body: { wrapper: this.animTag, } };
     return (
       <InnerTableContainer>
         <HeaderTitleTable small={true}>
           <Icon type={'plus-square'} theme={'filled'} onClick={this.handleAdd} />
-          <TextTitle small={true}>{'Cover Details'}</TextTitle>
+          <TextTitle small={true}>{'Contribution/Withdrawals'}</TextTitle>
           <DivideLine />
         </HeaderTitleTable>
-        <Table columns={columns} dataSource={dataSource} pagination={false} size={'small'} />
+        <Table columns={columns} dataSource={dataSource} pagination={false} components={components} size={'small'} />
       </InnerTableContainer>
     );
   }
 }
 
-export default CoverDetailsTable;
+export default ContributionWithdrawalsTable;
