+++ src/components/ClientDetailPage/assets/ExpandedAssetsRow.tsx
