+++ src/components/ClientDetailPage/insurance/ExpandedInsuranceRow.tsx
