+++ src/components/ClientDetailPage/liabilities/ExpandedLiabilitiesRow.tsx
