--- src/components/EntryTables/styled.ts
@@ -1,20 +1,0 @@
-
-import styled, { css } from 'styled-components';
-
-export const HeadingWrapper = styled.section`
-  &.default{
-    color: #252525;
-  }
-  &.subHeading{
-    color: #515c83;
-    .ant-typography{
-      margin: 35px 0px 0px 0px;
-    }
-  }
-  .ant-typography{
-    font-size: 34px;
-    color: inherit;
-    text-align: center;
-    margin: 35px 0px 35px 0px;
-  }
-`;
